+++ packages/plugin-code-block/web/src/editor.js
@@ -1,6 +1,6 @@
 import { createCodeBlockPlugin } from './createCodeBlockPlugin';
 import { CODE_BLOCK_TYPE } from './types';
-import { DEFAULTS } from './defaults';
+import { DEFAULTS, THEME as theme } from './defaults';
 
 export const pluginCodeBlock = (config = {}) => {
   return {
@@ -8,5 +8,6 @@ export const pluginCodeBlock = (config = {}) => {
     type: CODE_BLOCK_TYPE,
     createPlugin: createCodeBlockPlugin,
     ModalsMap: {},
+    theme,
   };
 };