+++ packages/ricos-common/web/src/themeStrategy/themes/viewer.ts
@@ -1,12 +1,20 @@
 /* eslint-disable camelcase */
-import { PaletteColors } from '../themeTypes';
+import { PaletteColors, ThemeUtils } from '../themeTypes';
 
-export default function viewer(colors: PaletteColors) {
-  const { actionColor } = colors;
+export default function viewer(colors: PaletteColors, utils: ThemeUtils) {
+  const { actionColor, textColor } = colors;
   return {
     quote: {
       'border-left-color': actionColor,
       'border-right-color': actionColor,
+    },
+    //rich-content-viewer.scss
+    code: {},
+    codeBlock: {},
+    editor: {
+      '& $code, $codeBlock': {
+        backgroundColor: utils.hexToRgbA(textColor, utils.isBright(textColor) ? 0.2 : 0.05),
+      },
     },
   };
 }