+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   - [#1303](https://github.com/wix-incubator/rich-content/pull/1303) actionColor fix for "more+" button
   - [#1304](https://github.com/wix-incubator/rich-content/pull/1304) textColor fix for plugin (+) button
   - [#1305](https://github.com/wix-incubator/rich-content/pull/1305) theme - toolbar white on white
+  - [#1306](https://github.com/wix-incubator/rich-content/pull/1306) theme - codeblock wiring
 </details>
 <hr/>